Brings in the build-ci, smatrix-hdf5, gtest-infra and python-bindings work
so that this branch merges into develop cleanly once those are in.

Three conflicts needed a decision. The dense matrix observable arrives here
in its old exit(1) form and is deleted on this branch, so the deletion wins.
matrixDataTypes.h was re-indented by the smatrix-hdf5 work while this branch
replaced an exit(1) with a throw; the new layout keeps the throw. The test
directory was renamed test -> tests, so testErrorHandling.cpp moves with it
and is registered in tests/CMakeLists.txt.

+++ core/krylovObservables.cpp
@@ -60,7 +60,7 @@ krylovBasicObservable::krylovBasicObservable(const std::string& name, std::vecto
 {
     expectationValues = new double[numSamples];
 
-    for (int i = 0; i != numSamples; i++) {
+    for (size_t i = 0; i != numSamples; i++) {
         expectationValues[i] = values[i];
     }
 }
@@ -124,7 +124,7 @@ std::complex<double> krylovSpMatrixObservable::expectation(std::complex<double>*
     {
         throw krylovError("Too many samples.");
     }
-	if (len != dim)
+	if (static_cast<size_t>(len) != dim)
 	{
 		throw krylovInvalidArgument("Incompatible dimensions");
 	}
@@ -165,7 +165,7 @@ std::complex<double> krylovVectorObservable::expectation(std::complex<double>* v
     {
         throw krylovError("Too many samples.");
     }
-	if (len != dim)
+	if (static_cast<size_t>(len) != dim)
 	{
 		throw krylovInvalidArgument("Incompatible dimensions");
 	}
@@ -188,7 +188,7 @@ krylovOutputObservable::krylovOutputObservable(const std::string& name, std::vec
 {
     this->numSamples = values.size();
     initializeResultArray(numSamples);
-    for (int i = 0; i != numSamples; i++) {
+    for (size_t i = 0; i != numSamples; i++) {
         expectationValues[i] = values[i];
     }
 }
